Add metrics endpoint
