+++ dev-env/config/config2.toml
@@ -25,6 +25,7 @@ generic_key_rate_limit = 2
 load_rate_limit = 60
 leader_rate_limit = 10
 add_task_rate_limit = 8
+metric_rate_limit = 5
 request_size_limit = 102400
 request_file_size_limit = 31457280
 wss_bittensor = "wss://entrypoint-finney.opentensor.ai:443"
