+++ src/common/crypto_provider.rs
