+++ src/common/mod.rs
@@ -1,5 +1,5 @@
 pub mod cert;
-pub mod crypto;
+pub mod crypto_provider;
 pub mod log;
 pub mod queue;
 pub mod resolve;
